pyproject.toml
[project]
name = "did-annuaire-be1"
version = "0.1.0"
description = "DID Annuaire - Decentralized identifier application"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3",
    "celery",
    "dj-database-url",
    "django",
    "django-celery-beat",
    "django-celery-results",
    "django-environ",
    "django-extensions",
    "django-filter",
    "django-ninja",
    "django-ninja-jwt[crypto]",
    "django-structlog[celery]",
    "gunicorn",
    "hvac",
    "jsonschema>=4.26.0",
    "psycopg[c]",
    "pyld",
    "pynacl",
    "pyotp",
    "qrcode[pil]",
    "sentry-sdk[django]",
]

[dependency-groups]
dev = [
    "django-debug-toolbar",
    "rav",
    "ruff",
]


